Create PowerOfTwo
In this method, repeatedly divide the number by 2. At some point, if the number becomes odd, it means that the number is not in the power of 2. Otherwise, if the number becomes 1 in the end, means the number is a power of 2.
e.g. 1. 32 -->32/2=16 --->16/2=8 -->8/2=4 --> 4/2=2 -->2/2=1, therefore 32 is a power of two.
       2. 28---->28/2=14 -->14/2=7, the number becomes odd, it means it is not a power of 2.